Movement Overhaul (pt. 2)
Ramps can now be supported. Some slight slippage on ramps still needs to be solved. Stairs unfortunately will not work with this solution at all.

I still want to look into other solutions. A complete rework of movement may be upcoming so that ramps and stairs will work in a cleaner way